ga-bq-pipeline/requirements.txt
google-cloud-storage==1.29.0
google-cloud-bigquery==1.25.0
google-cloud-bigquery-storage==0.6.0
jsonlines==1.2.0
urllib3==1.25.9
google-api-python-client==1.9.3
google-api-core==1.20.0
google-auth==1.17.2
pandas==0.25.3
pyarrow==0.16.0
PyYAML==5.3.1
ua-parser==0.8.0
user-agents==2.1
apache-airflow==1.10.5